requirements.txt
tweepy == 4.12.0
pandas ==1.3.5
python-decouple == 3.6
sqlalchemy == 1.4
psycopg2-binary == 2.9.5
